+++ x/oracle/types/event.pb.go
@@ -81,12 +81,11 @@ func (m *EventExchangeRateUpdate) GetTimestampMs() int64 {
 
 // Emitted when a valoper delegates oracle voting rights to a feeder address.
 type EventDelegateFeederConsent struct {
+	// Validator is the Bech32 address that is delegating voting rights.
+	Validator string `protobuf:"bytes,1,opt,name=validator,proto3" json:"validator,omitempty"`
 	// Feeder is the delegate or representative that will be able to send
 	// vote and prevote transaction messages.
-	Feeder string `protobuf:"bytes,1,opt,name=feeder,proto3" json:"feeder,omitempty"`
-	// Operator is the delegator, a valdiator operator that is delegating
-	// voting rights.
-	Operator string `protobuf:"bytes,2,opt,name=operator,proto3" json:"operator,omitempty"`
+	Feeder string `protobuf:"bytes,2,opt,name=feeder,proto3" json:"feeder,omitempty"`
 }
 
 func (m *EventDelegateFeederConsent) Reset()         { *m = EventDelegateFeederConsent{} }
@@ -122,6 +121,13 @@ func (m *EventDelegateFeederConsent) XXX_DiscardUnknown() {
 
 var xxx_messageInfo_EventDelegateFeederConsent proto.InternalMessageInfo
 
+func (m *EventDelegateFeederConsent) GetValidator() string {
+	if m != nil {
+		return m.Validator
+	}
+	return ""
+}
+
 func (m *EventDelegateFeederConsent) GetFeeder() string {
 	if m != nil {
 		return m.Feeder
@@ -129,9 +135,123 @@ func (m *EventDelegateFeederConsent) GetFeeder() string {
 	return ""
 }
 
-func (m *EventDelegateFeederConsent) GetOperator() string {
+// Emitted by MsgAggregateExchangeVote when an aggregate vote is added to state
+type EventAggregateVote struct {
+	// Validator is the Bech32 address to which the vote will be credited.
+	Validator string `protobuf:"bytes,1,opt,name=validator,proto3" json:"validator,omitempty"`
+	// Feeder is the delegate or representative that will sends vote and prevote
+	// transaction messages on behalf of the voting validator.
+	Feeder        string `protobuf:"bytes,2,opt,name=feeder,proto3" json:"feeder,omitempty"`
+	ExchangeRates string `protobuf:"bytes,3,opt,name=exchange_rates,json=exchangeRates,proto3" json:"exchange_rates,omitempty" yaml:"exchange_rates"`
+}
+
+func (m *EventAggregateVote) Reset()         { *m = EventAggregateVote{} }
+func (m *EventAggregateVote) String() string { return proto.CompactTextString(m) }
+func (*EventAggregateVote) ProtoMessage()    {}
+func (*EventAggregateVote) Descriptor() ([]byte, []int) {
+	return fileDescriptor_94ec441b793fc0ea, []int{2}
+}
+func (m *EventAggregateVote) XXX_Unmarshal(b []byte) error {
+	return m.Unmarshal(b)
+}
+func (m *EventAggregateVote) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
+	if deterministic {
+		return xxx_messageInfo_EventAggregateVote.Marshal(b, m, deterministic)
+	} else {
+		b = b[:cap(b)]
+		n, err := m.MarshalToSizedBuffer(b)
+		if err != nil {
+			return nil, err
+		}
+		return b[:n], nil
+	}
+}
+func (m *EventAggregateVote) XXX_Merge(src proto.Message) {
+	xxx_messageInfo_EventAggregateVote.Merge(m, src)
+}
+func (m *EventAggregateVote) XXX_Size() int {
+	return m.Size()
+}
+func (m *EventAggregateVote) XXX_DiscardUnknown() {
+	xxx_messageInfo_EventAggregateVote.DiscardUnknown(m)
+}
+
+var xxx_messageInfo_EventAggregateVote proto.InternalMessageInfo
+
+func (m *EventAggregateVote) GetValidator() string {
 	if m != nil {
-		return m.Operator
+		return m.Validator
+	}
+	return ""
+}
+
+func (m *EventAggregateVote) GetFeeder() string {
+	if m != nil {
+		return m.Feeder
+	}
+	return ""
+}
+
+func (m *EventAggregateVote) GetExchangeRates() string {
+	if m != nil {
+		return m.ExchangeRates
+	}
+	return ""
+}
+
+// Emitted by MsgAggregateExchangePrevote when an aggregate prevote is added
+// to state
+type EventAggregatePrevote struct {
+	// Validator is the Bech32 address to which the vote will be credited.
+	Validator string `protobuf:"bytes,1,opt,name=validator,proto3" json:"validator,omitempty"`
+	// Feeder is the delegate or representative that will sends vote and prevote
+	// transaction messages on behalf of the voting validator.
+	Feeder string `protobuf:"bytes,2,opt,name=feeder,proto3" json:"feeder,omitempty"`
+}
+
+func (m *EventAggregatePrevote) Reset()         { *m = EventAggregatePrevote{} }
+func (m *EventAggregatePrevote) String() string { return proto.CompactTextString(m) }
+func (*EventAggregatePrevote) ProtoMessage()    {}
+func (*EventAggregatePrevote) Descriptor() ([]byte, []int) {
+	return fileDescriptor_94ec441b793fc0ea, []int{3}
+}
+func (m *EventAggregatePrevote) XXX_Unmarshal(b []byte) error {
+	return m.Unmarshal(b)
+}
+func (m *EventAggregatePrevote) XXX_Marshal(b []byte, deterministic bool) ([]byte, error) {
+	if deterministic {
+		return xxx_messageInfo_EventAggregatePrevote.Marshal(b, m, deterministic)
+	} else {
+		b = b[:cap(b)]
+		n, err := m.MarshalToSizedBuffer(b)
+		if err != nil {
+			return nil, err
+		}
+		return b[:n], nil
+	}
+}
+func (m *EventAggregatePrevote) XXX_Merge(src proto.Message) {
+	xxx_messageInfo_EventAggregatePrevote.Merge(m, src)
+}
+func (m *EventAggregatePrevote) XXX_Size() int {
+	return m.Size()
+}
+func (m *EventAggregatePrevote) XXX_DiscardUnknown() {
+	xxx_messageInfo_EventAggregatePrevote.DiscardUnknown(m)
+}
+
+var xxx_messageInfo_EventAggregatePrevote proto.InternalMessageInfo
+
+func (m *EventAggregatePrevote) GetValidator() string {
+	if m != nil {
+		return m.Validator
+	}
+	return ""
+}
+
+func (m *EventAggregatePrevote) GetFeeder() string {
+	if m != nil {
+		return m.Feeder
 	}
 	return ""
 }
@@ -139,33 +259,39 @@ func (m *EventDelegateFeederConsent) GetOperator() string {
 func init() {
 	proto.RegisterType((*EventExchangeRateUpdate)(nil), "nibiru.oracle.v1.EventExchangeRateUpdate")
 	proto.RegisterType((*EventDelegateFeederConsent)(nil), "nibiru.oracle.v1.EventDelegateFeederConsent")
+	proto.RegisterType((*EventAggregateVote)(nil), "nibiru.oracle.v1.EventAggregateVote")
+	proto.RegisterType((*EventAggregatePrevote)(nil), "nibiru.oracle.v1.EventAggregatePrevote")
 }
 
 func init() { proto.RegisterFile("nibiru/oracle/v1/event.proto", fileDescriptor_94ec441b793fc0ea) }
 
 var fileDescriptor_94ec441b793fc0ea = []byte{
-	// 332 bytes of a gzipped FileDescriptorProto
-	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0x4c, 0x91, 0xcf, 0x4e, 0xfa, 0x40,
-	0x10, 0xc7, 0xdb, 0x1f, 0xbf, 0x10, 0x5d, 0x31, 0x31, 0x8d, 0x51, 0xd2, 0x90, 0x82, 0x1c, 0x0c,
-	0x07, 0xed, 0x86, 0xf8, 0x06, 0xfc, 0xbb, 0x69, 0x4c, 0x8d, 0x17, 0x2f, 0x64, 0x29, 0x63, 0xd9,
-	0x48, 0x77, 0x9a, 0xdd, 0x81, 0xe0, 0x5b, 0xf8, 0x12, 0xbe, 0x0b, 0x47, 0x8e, 0xc6, 0x03, 0x31,
-	0xf0, 0x22, 0x86, 0x6d, 0x25, 0x9c, 0x76, 0x66, 0x3e, 0x93, 0x9d, 0xef, 0xcc, 0x97, 0xd5, 0x94,
-	0x1c, 0x49, 0x3d, 0xe3, 0xa8, 0x45, 0x3c, 0x05, 0x3e, 0x6f, 0x73, 0x98, 0x83, 0xa2, 0x30, 0xd3,
-	0x48, 0xe8, 0x9d, 0xe5, 0x34, 0xcc, 0x69, 0x38, 0x6f, 0xfb, 0xe7, 0x09, 0x26, 0x68, 0x21, 0xdf,
-	0x45, 0x79, 0x9f, 0x5f, 0x4b, 0x10, 0x93, 0x29, 0x70, 0x91, 0x49, 0x2e, 0x94, 0x42, 0x12, 0x24,
-	0x51, 0x99, 0x9c, 0x36, 0x3f, 0x5d, 0x76, 0xd9, 0xdf, 0xfd, 0xda, 0x5f, 0xc4, 0x13, 0xa1, 0x12,
-	0x88, 0x04, 0xc1, 0x73, 0x36, 0x16, 0x04, 0x9e, 0xc7, 0xfe, 0x67, 0x42, 0xea, 0xaa, 0xdb, 0x70,
-	0x5b, 0xc7, 0x91, 0x8d, 0xbd, 0x27, 0x76, 0x0a, 0x45, 0xe7, 0x50, 0x0b, 0x82, 0xea, 0xbf, 0x1d,
-	0xec, 0x84, 0xcb, 0x75, 0xdd, 0xf9, 0x5e, 0xd7, 0xaf, 0x13, 0x49, 0x93, 0xd9, 0x28, 0x8c, 0x31,
-	0xe5, 0x31, 0x9a, 0x14, 0x4d, 0xf1, 0xdc, 0x9a, 0xf1, 0x1b, 0xa7, 0xf7, 0x0c, 0x4c, 0xd8, 0x83,
-	0x38, 0xaa, 0xc0, 0xc1, 0x38, 0xef, 0x8a, 0x55, 0x48, 0xa6, 0x60, 0x48, 0xa4, 0xd9, 0x30, 0x35,
-	0xd5, 0x52, 0xc3, 0x6d, 0x95, 0xa2, 0x93, 0x7d, 0xed, 0xde, 0x34, 0x1f, 0x99, 0x6f, 0x65, 0xf6,
-	0x60, 0x0a, 0x89, 0x20, 0x18, 0x00, 0x8c, 0x41, 0x77, 0x51, 0x19, 0x50, 0xe4, 0x5d, 0xb0, 0xf2,
-	0xab, 0x2d, 0x14, 0x5a, 0x8b, 0xcc, 0xf3, 0xd9, 0x11, 0x66, 0xa0, 0x05, 0xa1, 0xce, 0x85, 0x46,
-	0xfb, 0xbc, 0x33, 0x58, 0x6e, 0x02, 0x77, 0xb5, 0x09, 0xdc, 0x9f, 0x4d, 0xe0, 0x7e, 0x6c, 0x03,
-	0x67, 0xb5, 0x0d, 0x9c, 0xaf, 0x6d, 0xe0, 0xbc, 0xdc, 0x1c, 0x2c, 0xf1, 0x60, 0x8f, 0xdc, 0x9d,
-	0x08, 0xa9, 0x78, 0x61, 0xc7, 0xe2, 0xcf, 0x10, 0xbb, 0xce, 0xa8, 0x6c, 0x0f, 0x79, 0xf7, 0x1b,
-	0x00, 0x00, 0xff, 0xff, 0x9a, 0x6c, 0xc7, 0x14, 0xae, 0x01, 0x00, 0x00,
+	// 396 bytes of a gzipped FileDescriptorProto
+	0x1f, 0x8b, 0x08, 0x00, 0x00, 0x00, 0x00, 0x00, 0x02, 0xff, 0x9c, 0x52, 0xcd, 0xae, 0x12, 0x31,
+	0x14, 0x9e, 0x7a, 0xcd, 0x4d, 0xa8, 0xf7, 0x1a, 0xd3, 0x88, 0x22, 0x21, 0x03, 0xce, 0xc2, 0xb0,
+	0xd0, 0x69, 0x88, 0x3b, 0x57, 0xca, 0xdf, 0x0e, 0x63, 0xc6, 0xe8, 0xc2, 0x0d, 0x29, 0x33, 0xc7,
+	0xd2, 0x38, 0xd3, 0x4e, 0xda, 0x32, 0x81, 0x77, 0x70, 0xe1, 0x4b, 0xf8, 0x2e, 0x2c, 0x59, 0x1a,
+	0x17, 0xc4, 0xc0, 0x1b, 0xf8, 0x04, 0x86, 0xce, 0xe0, 0x85, 0x2d, 0xab, 0x9e, 0x9e, 0xef, 0xcb,
+	0xd7, 0xaf, 0xe7, 0x7c, 0xb8, 0x25, 0xc5, 0x4c, 0xe8, 0x05, 0x55, 0x9a, 0xc5, 0x29, 0xd0, 0xa2,
+	0x47, 0xa1, 0x00, 0x69, 0xc3, 0x5c, 0x2b, 0xab, 0xc8, 0xa3, 0x12, 0x0d, 0x4b, 0x34, 0x2c, 0x7a,
+	0xcd, 0xc7, 0x5c, 0x71, 0xe5, 0x40, 0x7a, 0xa8, 0x4a, 0x5e, 0xb3, 0xc5, 0x95, 0xe2, 0x29, 0x50,
+	0x96, 0x0b, 0xca, 0xa4, 0x54, 0x96, 0x59, 0xa1, 0xa4, 0x29, 0xd1, 0xe0, 0x27, 0xc2, 0x4f, 0x47,
+	0x07, 0xd5, 0xd1, 0x32, 0x9e, 0x33, 0xc9, 0x21, 0x62, 0x16, 0x3e, 0xe5, 0x09, 0xb3, 0x40, 0x08,
+	0xbe, 0x9f, 0x33, 0xa1, 0x1b, 0xa8, 0x83, 0xba, 0xb5, 0xc8, 0xd5, 0xe4, 0x23, 0xbe, 0x85, 0x8a,
+	0x39, 0xd5, 0xcc, 0x42, 0xe3, 0xde, 0x01, 0xec, 0x87, 0xeb, 0x6d, 0xdb, 0xfb, 0xbd, 0x6d, 0xbf,
+	0xe0, 0xc2, 0xce, 0x17, 0xb3, 0x30, 0x56, 0x19, 0x8d, 0x95, 0xc9, 0x94, 0xa9, 0x8e, 0x57, 0x26,
+	0xf9, 0x46, 0xed, 0x2a, 0x07, 0x13, 0x0e, 0x21, 0x8e, 0x6e, 0xe0, 0xe4, 0x39, 0xf2, 0x1c, 0xdf,
+	0x58, 0x91, 0x81, 0xb1, 0x2c, 0xcb, 0xa7, 0x99, 0x69, 0x5c, 0x75, 0x50, 0xf7, 0x2a, 0x7a, 0xf0,
+	0xbf, 0x37, 0x31, 0x41, 0x84, 0x9b, 0xce, 0xe6, 0x10, 0x52, 0xe0, 0xcc, 0xc2, 0x18, 0x20, 0x01,
+	0x3d, 0x50, 0xd2, 0x80, 0xb4, 0xa4, 0x85, 0x6b, 0x05, 0x4b, 0x45, 0xc2, 0xac, 0x3a, 0xda, 0xbd,
+	0x6b, 0x90, 0x27, 0xf8, 0xfa, 0xab, 0xa3, 0x97, 0x66, 0xa3, 0xea, 0x16, 0x7c, 0x47, 0x98, 0x38,
+	0xd1, 0x77, 0x9c, 0x6b, 0xa7, 0xfa, 0x59, 0x59, 0xb8, 0x4c, 0x8c, 0xbc, 0xc5, 0x0f, 0xcf, 0x06,
+	0x53, 0xfe, 0xa2, 0xd6, 0x7f, 0xf6, 0x77, 0xdb, 0xae, 0xaf, 0x58, 0x96, 0xbe, 0x09, 0xce, 0xf1,
+	0x20, 0xba, 0x3d, 0x1d, 0x82, 0x09, 0x26, 0xb8, 0x7e, 0xee, 0xe6, 0x83, 0x86, 0xe2, 0x62, 0x43,
+	0xfd, 0xf1, 0x7a, 0xe7, 0xa3, 0xcd, 0xce, 0x47, 0x7f, 0x76, 0x3e, 0xfa, 0xb1, 0xf7, 0xbd, 0xcd,
+	0xde, 0xf7, 0x7e, 0xed, 0x7d, 0xef, 0xcb, 0xcb, 0x93, 0x25, 0xbd, 0x77, 0x21, 0x1a, 0xcc, 0x99,
+	0x90, 0xb4, 0x8a, 0xdb, 0xf2, 0x18, 0x38, 0xb7, 0xae, 0xd9, 0xb5, 0x0b, 0xca, 0xeb, 0x7f, 0x01,
+	0x00, 0x00, 0xff, 0xff, 0xc8, 0x26, 0xa5, 0x2c, 0x8e, 0x02, 0x00, 0x00,
 }
 
 func (m *EventExchangeRateUpdate) Marshal() (dAtA []byte, err error) {
@@ -233,17 +359,98 @@ func (m *EventDelegateFeederConsent) MarshalToSizedBuffer(dAtA []byte) (int, err
 	_ = i
 	var l int
 	_ = l
-	if len(m.Operator) > 0 {
-		i -= len(m.Operator)
-		copy(dAtA[i:], m.Operator)
-		i = encodeVarintEvent(dAtA, i, uint64(len(m.Operator)))
+	if len(m.Feeder) > 0 {
+		i -= len(m.Feeder)
+		copy(dAtA[i:], m.Feeder)
+		i = encodeVarintEvent(dAtA, i, uint64(len(m.Feeder)))
 		i--
 		dAtA[i] = 0x12
+	}
+	if len(m.Validator) > 0 {
+		i -= len(m.Validator)
+		copy(dAtA[i:], m.Validator)
+		i = encodeVarintEvent(dAtA, i, uint64(len(m.Validator)))
+		i--
+		dAtA[i] = 0xa
+	}
+	return len(dAtA) - i, nil
+}
+
+func (m *EventAggregateVote) Marshal() (dAtA []byte, err error) {
+	size := m.Size()
+	dAtA = make([]byte, size)
+	n, err := m.MarshalToSizedBuffer(dAtA[:size])
+	if err != nil {
+		return nil, err
+	}
+	return dAtA[:n], nil
+}
+
+func (m *EventAggregateVote) MarshalTo(dAtA []byte) (int, error) {
+	size := m.Size()
+	return m.MarshalToSizedBuffer(dAtA[:size])
+}
+
+func (m *EventAggregateVote) MarshalToSizedBuffer(dAtA []byte) (int, error) {
+	i := len(dAtA)
+	_ = i
+	var l int
+	_ = l
+	if len(m.ExchangeRates) > 0 {
+		i -= len(m.ExchangeRates)
+		copy(dAtA[i:], m.ExchangeRates)
+		i = encodeVarintEvent(dAtA, i, uint64(len(m.ExchangeRates)))
+		i--
+		dAtA[i] = 0x1a
 	}
 	if len(m.Feeder) > 0 {
 		i -= len(m.Feeder)
 		copy(dAtA[i:], m.Feeder)
 		i = encodeVarintEvent(dAtA, i, uint64(len(m.Feeder)))
+		i--
+		dAtA[i] = 0x12
+	}
+	if len(m.Validator) > 0 {
+		i -= len(m.Validator)
+		copy(dAtA[i:], m.Validator)
+		i = encodeVarintEvent(dAtA, i, uint64(len(m.Validator)))
+		i--
+		dAtA[i] = 0xa
+	}
+	return len(dAtA) - i, nil
+}
+
+func (m *EventAggregatePrevote) Marshal() (dAtA []byte, err error) {
+	size := m.Size()
+	dAtA = make([]byte, size)
+	n, err := m.MarshalToSizedBuffer(dAtA[:size])
+	if err != nil {
+		return nil, err
+	}
+	return dAtA[:n], nil
+}
+
+func (m *EventAggregatePrevote) MarshalTo(dAtA []byte) (int, error) {
+	size := m.Size()
+	return m.MarshalToSizedBuffer(dAtA[:size])
+}
+
+func (m *EventAggregatePrevote) MarshalToSizedBuffer(dAtA []byte) (int, error) {
+	i := len(dAtA)
+	_ = i
+	var l int
+	_ = l
+	if len(m.Feeder) > 0 {
+		i -= len(m.Feeder)
+		copy(dAtA[i:], m.Feeder)
+		i = encodeVarintEvent(dAtA, i, uint64(len(m.Feeder)))
+		i--
+		dAtA[i] = 0x12
+	}
+	if len(m.Validator) > 0 {
+		i -= len(m.Validator)
+		copy(dAtA[i:], m.Validator)
+		i = encodeVarintEvent(dAtA, i, uint64(len(m.Validator)))
 		i--
 		dAtA[i] = 0xa
 	}
@@ -285,11 +492,49 @@ func (m *EventDelegateFeederConsent) Size() (n int) {
 	}
 	var l int
 	_ = l
+	l = len(m.Validator)
+	if l > 0 {
+		n += 1 + l + sovEvent(uint64(l))
+	}
 	l = len(m.Feeder)
 	if l > 0 {
 		n += 1 + l + sovEvent(uint64(l))
 	}
-	l = len(m.Operator)
+	return n
+}
+
+func (m *EventAggregateVote) Size() (n int) {
+	if m == nil {
+		return 0
+	}
+	var l int
+	_ = l
+	l = len(m.Validator)
+	if l > 0 {
+		n += 1 + l + sovEvent(uint64(l))
+	}
+	l = len(m.Feeder)
+	if l > 0 {
+		n += 1 + l + sovEvent(uint64(l))
+	}
+	l = len(m.ExchangeRates)
+	if l > 0 {
+		n += 1 + l + sovEvent(uint64(l))
+	}
+	return n
+}
+
+func (m *EventAggregatePrevote) Size() (n int) {
+	if m == nil {
+		return 0
+	}
+	var l int
+	_ = l
+	l = len(m.Validator)
+	if l > 0 {
+		n += 1 + l + sovEvent(uint64(l))
+	}
+	l = len(m.Feeder)
 	if l > 0 {
 		n += 1 + l + sovEvent(uint64(l))
 	}
@@ -468,6 +713,38 @@ func (m *EventDelegateFeederConsent) Unmarshal(dAtA []byte) error {
 		switch fieldNum {
 		case 1:
 			if wireType != 2 {
+				return fmt.Errorf("proto: wrong wireType = %d for field Validator", wireType)
+			}
+			var stringLen uint64
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowEvent
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				stringLen |= uint64(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+			intStringLen := int(stringLen)
+			if intStringLen < 0 {
+				return ErrInvalidLengthEvent
+			}
+			postIndex := iNdEx + intStringLen
+			if postIndex < 0 {
+				return ErrInvalidLengthEvent
+			}
+			if postIndex > l {
+				return io.ErrUnexpectedEOF
+			}
+			m.Validator = string(dAtA[iNdEx:postIndex])
+			iNdEx = postIndex
+		case 2:
+			if wireType != 2 {
 				return fmt.Errorf("proto: wrong wireType = %d for field Feeder", wireType)
 			}
 			var stringLen uint64
@@ -498,9 +775,59 @@ func (m *EventDelegateFeederConsent) Unmarshal(dAtA []byte) error {
 			}
 			m.Feeder = string(dAtA[iNdEx:postIndex])
 			iNdEx = postIndex
-		case 2:
+		default:
+			iNdEx = preIndex
+			skippy, err := skipEvent(dAtA[iNdEx:])
+			if err != nil {
+				return err
+			}
+			if (skippy < 0) || (iNdEx+skippy) < 0 {
+				return ErrInvalidLengthEvent
+			}
+			if (iNdEx + skippy) > l {
+				return io.ErrUnexpectedEOF
+			}
+			iNdEx += skippy
+		}
+	}
+
+	if iNdEx > l {
+		return io.ErrUnexpectedEOF
+	}
+	return nil
+}
+func (m *EventAggregateVote) Unmarshal(dAtA []byte) error {
+	l := len(dAtA)
+	iNdEx := 0
+	for iNdEx < l {
+		preIndex := iNdEx
+		var wire uint64
+		for shift := uint(0); ; shift += 7 {
+			if shift >= 64 {
+				return ErrIntOverflowEvent
+			}
+			if iNdEx >= l {
+				return io.ErrUnexpectedEOF
+			}
+			b := dAtA[iNdEx]
+			iNdEx++
+			wire |= uint64(b&0x7F) << shift
+			if b < 0x80 {
+				break
+			}
+		}
+		fieldNum := int32(wire >> 3)
+		wireType := int(wire & 0x7)
+		if wireType == 4 {
+			return fmt.Errorf("proto: EventAggregateVote: wiretype end group for non-group")
+		}
+		if fieldNum <= 0 {
+			return fmt.Errorf("proto: EventAggregateVote: illegal tag %d (wire type %d)", fieldNum, wire)
+		}
+		switch fieldNum {
+		case 1:
 			if wireType != 2 {
-				return fmt.Errorf("proto: wrong wireType = %d for field Operator", wireType)
+				return fmt.Errorf("proto: wrong wireType = %d for field Validator", wireType)
 			}
 			var stringLen uint64
 			for shift := uint(0); ; shift += 7 {
@@ -528,7 +855,185 @@ func (m *EventDelegateFeederConsent) Unmarshal(dAtA []byte) error {
 			if postIndex > l {
 				return io.ErrUnexpectedEOF
 			}
-			m.Operator = string(dAtA[iNdEx:postIndex])
+			m.Validator = string(dAtA[iNdEx:postIndex])
+			iNdEx = postIndex
+		case 2:
+			if wireType != 2 {
+				return fmt.Errorf("proto: wrong wireType = %d for field Feeder", wireType)
+			}
+			var stringLen uint64
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowEvent
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				stringLen |= uint64(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+			intStringLen := int(stringLen)
+			if intStringLen < 0 {
+				return ErrInvalidLengthEvent
+			}
+			postIndex := iNdEx + intStringLen
+			if postIndex < 0 {
+				return ErrInvalidLengthEvent
+			}
+			if postIndex > l {
+				return io.ErrUnexpectedEOF
+			}
+			m.Feeder = string(dAtA[iNdEx:postIndex])
+			iNdEx = postIndex
+		case 3:
+			if wireType != 2 {
+				return fmt.Errorf("proto: wrong wireType = %d for field ExchangeRates", wireType)
+			}
+			var stringLen uint64
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowEvent
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				stringLen |= uint64(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+			intStringLen := int(stringLen)
+			if intStringLen < 0 {
+				return ErrInvalidLengthEvent
+			}
+			postIndex := iNdEx + intStringLen
+			if postIndex < 0 {
+				return ErrInvalidLengthEvent
+			}
+			if postIndex > l {
+				return io.ErrUnexpectedEOF
+			}
+			m.ExchangeRates = string(dAtA[iNdEx:postIndex])
+			iNdEx = postIndex
+		default:
+			iNdEx = preIndex
+			skippy, err := skipEvent(dAtA[iNdEx:])
+			if err != nil {
+				return err
+			}
+			if (skippy < 0) || (iNdEx+skippy) < 0 {
+				return ErrInvalidLengthEvent
+			}
+			if (iNdEx + skippy) > l {
+				return io.ErrUnexpectedEOF
+			}
+			iNdEx += skippy
+		}
+	}
+
+	if iNdEx > l {
+		return io.ErrUnexpectedEOF
+	}
+	return nil
+}
+func (m *EventAggregatePrevote) Unmarshal(dAtA []byte) error {
+	l := len(dAtA)
+	iNdEx := 0
+	for iNdEx < l {
+		preIndex := iNdEx
+		var wire uint64
+		for shift := uint(0); ; shift += 7 {
+			if shift >= 64 {
+				return ErrIntOverflowEvent
+			}
+			if iNdEx >= l {
+				return io.ErrUnexpectedEOF
+			}
+			b := dAtA[iNdEx]
+			iNdEx++
+			wire |= uint64(b&0x7F) << shift
+			if b < 0x80 {
+				break
+			}
+		}
+		fieldNum := int32(wire >> 3)
+		wireType := int(wire & 0x7)
+		if wireType == 4 {
+			return fmt.Errorf("proto: EventAggregatePrevote: wiretype end group for non-group")
+		}
+		if fieldNum <= 0 {
+			return fmt.Errorf("proto: EventAggregatePrevote: illegal tag %d (wire type %d)", fieldNum, wire)
+		}
+		switch fieldNum {
+		case 1:
+			if wireType != 2 {
+				return fmt.Errorf("proto: wrong wireType = %d for field Validator", wireType)
+			}
+			var stringLen uint64
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowEvent
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				stringLen |= uint64(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+			intStringLen := int(stringLen)
+			if intStringLen < 0 {
+				return ErrInvalidLengthEvent
+			}
+			postIndex := iNdEx + intStringLen
+			if postIndex < 0 {
+				return ErrInvalidLengthEvent
+			}
+			if postIndex > l {
+				return io.ErrUnexpectedEOF
+			}
+			m.Validator = string(dAtA[iNdEx:postIndex])
+			iNdEx = postIndex
+		case 2:
+			if wireType != 2 {
+				return fmt.Errorf("proto: wrong wireType = %d for field Feeder", wireType)
+			}
+			var stringLen uint64
+			for shift := uint(0); ; shift += 7 {
+				if shift >= 64 {
+					return ErrIntOverflowEvent
+				}
+				if iNdEx >= l {
+					return io.ErrUnexpectedEOF
+				}
+				b := dAtA[iNdEx]
+				iNdEx++
+				stringLen |= uint64(b&0x7F) << shift
+				if b < 0x80 {
+					break
+				}
+			}
+			intStringLen := int(stringLen)
+			if intStringLen < 0 {
+				return ErrInvalidLengthEvent
+			}
+			postIndex := iNdEx + intStringLen
+			if postIndex < 0 {
+				return ErrInvalidLengthEvent
+			}
+			if postIndex > l {
+				return io.ErrUnexpectedEOF
+			}
+			m.Feeder = string(dAtA[iNdEx:postIndex])
 			iNdEx = postIndex
 		default:
 			iNdEx = preIndex