+++ CHANGELOG.md
@@ -56,6 +56,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### State Machine Breaking
 
 * [#1473](https://github.com/NibiruChain/nibiru/pull/1473) - refactor(perp)!: rename `OpenPosition` to `MarketOrder`
+* [#xxx](https://github.com/NibiruChain/nibiru/pull/xxx) - refactor(oracle)!: Move away from deprecated events to typed events in x/oracle
 
 ### Dependencies